vertex_flow/plugins/crypto_trading/requirements.txt
requests>=2.28.0
websocket-client>=1.6.0
pandas>=2.0.0
numpy>=1.24.0
ta-lib>=0.4.0
python-dotenv>=1.0.0
python-binance>=1.0.16
okx>=1.0.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
cryptography>=3.4.8
pydantic>=1.10.0